(PC-33428) feat(NC): move formatCurrencyFromCents in shared/currency
