+++ src/features/birthdayNotifications/pages/RecreditBirthdayNotification.tsx
@@ -7,8 +7,8 @@ import { useResetRecreditAmountToShow } from 'features/profile/api/useResetRecre
 import { useAppStateChange } from 'libs/appState'
 import { useGetPacificFrancToEuroRate } from 'libs/firebase/firestore/exchangeRates/useGetPacificFrancToEuroRate'
 import LottieView from 'libs/lottie'
-import { formatCurrencyFromCents } from 'libs/parsers/formatCurrencyFromCents'
 import { storage } from 'libs/storage'
+import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
 import { useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { getAge } from 'shared/user/getAge'
 import { useAvailableCredit } from 'shared/user/useAvailableCredit'
