+++ src/features/bookOffer/components/BookingDetails.tsx
@@ -25,8 +25,8 @@ import { formatFullAddressStartsWithPostalCode } from 'libs/address/useFormatFul
 import { useGetPacificFrancToEuroRate } from 'libs/firebase/firestore/exchangeRates/useGetPacificFrancToEuroRate'
 import { useIsFalseWithDelay } from 'libs/hooks/useIsFalseWithDelay'
 import { useLocation } from 'libs/location'
-import { formatCurrencyFromCents } from 'libs/parsers/formatCurrencyFromCents'
 import { useSubcategoriesMapping } from 'libs/subcategories'
+import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
 import { useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
 import { InfoBanner } from 'ui/components/banners/InfoBanner'
