+++ src/features/bookings/components/CancelBookingModal.tsx
@@ -14,8 +14,8 @@ import { isUserExBeneficiary } from 'features/profile/helpers/isUserExBeneficiar
 import { analytics } from 'libs/analytics'
 import { useGetPacificFrancToEuroRate } from 'libs/firebase/firestore/exchangeRates/useGetPacificFrancToEuroRate'
 import { useNetInfoContext } from 'libs/network/NetInfoWrapper'
-import { formatCurrencyFromCents } from 'libs/parsers/formatCurrencyFromCents'
 import { convertCentsToEuros } from 'libs/parsers/pricesConversion'
+import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
 import { Currency, useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonTertiaryPrimary } from 'ui/components/buttons/ButtonTertiaryPrimary'
