+++ src/features/offer/components/MovieScreeningCalendar/useSelectedDateScreenings.ts
@@ -8,7 +8,7 @@ import {
   MovieScreeningUserData,
 } from 'features/offer/components/MovieScreeningCalendar/types'
 import { useGetPacificFrancToEuroRate } from 'libs/firebase/firestore/exchangeRates/useGetPacificFrancToEuroRate'
-import { formatCurrencyFromCents } from 'libs/parsers/formatCurrencyFromCents'
+import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
 import { Currency, useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { EventCardProps } from 'ui/components/eventCard/EventCard'
