+++ src/features/profile/components/CreditInfo/CreditInfo.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components/native'
 import { DomainsCredit } from 'api/gen'
 import { CreditProgressBar } from 'features/profile/components/CreditInfo/CreditProgressBar'
 import { useGetPacificFrancToEuroRate } from 'libs/firebase/firestore/exchangeRates/useGetPacificFrancToEuroRate'
-import { formatCurrencyFromCents } from 'libs/parsers/formatCurrencyFromCents'
+import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
 import { useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { Spacer } from 'ui/theme'
 import { TypoDS } from 'ui/theme/designSystemTypographie'
